+++ cfdmod/analysis/__init__.py
@@ -1,6 +1,0 @@
-__all__ = [
-    "NormalizationParameters",
-    "InflowData",
-]
-
-from cfdmod.analysis.inflow import NormalizationParameters, InflowData