+++ cfdmod/analysis/inflow/__init__.py
@@ -1,8 +1,0 @@
-"""Inflow analysis module"""
-
-__all__ = [
-    "NormalizationParameters",
-    "InflowData",
-]
-
-from cfdmod.analysis.inflow.profile import NormalizationParameters, InflowData